fix: complete type import reorganization and lint error resolution
- Remove core type re-exports from web-api.ts to eliminate import confusion
- Update all imports to use correct type sources:
  - Core domain types (ThreadId, Task, etc.) from @/types/core
  - API serialization types (ApiSession, ApiAgent) from @/types/api
  - Event types from @/types/web-sse and @/types/web-events
- Fix all 137 lint errors caused by type import issues
- Separate API types properly use string representations of serialized types
- Maintain clean separation between domain types and wire format types

This sets up the foundation for implementing superjson for transparent
type serialization across client-server boundaries.
