+++ packages/web/app/api/sessions/[sessionId]/history/route.test.ts
@@ -9,7 +9,8 @@ import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest';
 import { NextRequest } from 'next/server';
 import { GET } from '@/app/api/sessions/[sessionId]/history/route';
 import { getSessionService } from '@/lib/server/session-service';
-import type { SessionEvent, ApiErrorResponse } from '@/types/api';
+import type { SessionEvent } from '@/types/web-sse';
+import type { ApiErrorResponse } from '@/types/api';
 import { asThreadId } from '@/lib/server/core-types';
 import { setupTestPersistence, teardownTestPersistence } from '~/test-utils/persistence-helper';
 import { Project } from '@/lib/server/lace-imports';
