+++ packages/web/app/api/threads/[threadId]/message/route.ts
@@ -4,7 +4,8 @@
 import { NextRequest, NextResponse } from 'next/server';
 import { randomUUID } from 'crypto';
 import { getSessionService } from '@/lib/server/session-service';
-import { MessageResponse, SessionEvent, ApiErrorResponse } from '@/types/api';
+import { MessageResponse, ApiErrorResponse } from '@/types/api';
+import type { SessionEvent } from '@/types/web-sse';
 import { asThreadId, type ThreadId } from '@/lib/server/core-types';
 import { EventStreamManager } from '@/lib/event-stream-manager';
 import { ThreadIdSchema, MessageRequestSchema } from '@/lib/validation/schemas';
