+++ packages/web/components/timeline/tool/task.tsx
@@ -19,7 +19,7 @@ import {
   faArrowRight
 } from '@fortawesome/free-solid-svg-icons';
 import type { ToolRenderer, ToolResult } from './types';
-import type { ToolAggregatedEventData } from '@/types/api';
+import type { ToolAggregatedEventData } from '@/types/web-events';
 import type { Task } from '@/types/core';
 import Badge from '@/components/ui/Badge';
 import InlineCode from '@/components/ui/InlineCode';
