+++ packages/web/components/timeline/tool/types.ts
@@ -2,7 +2,7 @@
 // ABOUTME: Defines interfaces and helper types for customizable tool display logic
 
 import type { ToolResult } from '@/lib/server/lace-imports';
-import type { ToolAggregatedEventData } from '@/types/api';
+import type { ToolAggregatedEventData } from '@/types/web-events';
 
 export interface ToolRenderer {
   getDisplayName?: (toolName: string, result?: ToolResult) => string;
