+++ packages/web/lib/task-metadata-realtime-integration.test.ts
@@ -3,7 +3,9 @@
 
 import { describe, it, expect } from 'vitest';
 import { convertSessionEventsToTimeline } from './timeline-converter';
-import type { SessionEvent, ApiAgent as AgentType, ToolResult, ThreadId } from '@/types/api';
+import type { SessionEvent } from '@/types/web-sse';
+import type { ApiAgent as AgentType } from '@/types/api';
+import type { ToolResult, ThreadId } from '@/types/core';
 
 describe('Task Metadata Event Format Bug', () => {
   it('should produce identical timeline results when TOOL_RESULT formats match', () => {
