+++ packages/web/lib/timeline-converter.ts
@@ -4,7 +4,10 @@
 // TODO: We should refactor to eliminate this conversion layer - either make TimelineEntry match SessionEvent
 // TODO: or standardize on one event format throughout the system to avoid this translation step.
 
-import type { SessionEvent, ApiAgent, ThreadId, ToolCallEventData } from '@/types/api';
+import type { ApiAgent } from '@/types/api';
+import type { SessionEvent } from '@/types/web-sse';
+import type { ToolCallEventData } from '@/types/web-events';
+import type { ThreadId } from '@/types/core';
 import type { ToolResult } from '@/lib/server/lace-imports';
 import type { TimelineEntry } from '@/types/design-system';
